pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.1",
    "discord-py>=2.6.4",
    "python-dotenv>=1.1.1",
    "supabase>=2.22.0",
    "httpx>=0.27.0",
]